+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for croud
 Unreleased
 ==========
 
+0.5.0 - 2018/01/22
+==================
+
 - Fix: Delegate all occurring error messages to the console output
 
 - Added `users roles remove` sub command that allows to remove a role from a